Initial commit after relocated server.js and package.json to root directory
