+++ .gitignore
@@ -1,4 +1,5 @@
 server/node_modules
+node_modules
 server/config/secret.json
 .DS_Store
 2018 Gala Auction Item Tracker - 2018 Auction Item Tracker.csv
