+++ package.json
@@ -1,0 +1,37 @@
+{
+  "name": "POTR",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "start": "sudo NODE_ENV=production node server.js",
+    "dev": "open http://localhost:8000/users/login & nodemon server.js",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "bcrypt-nodejs": "0.0.3",
+    "body-parser": "^1.17.2",
+    "csvtojson": "^2.0.6",
+    "dateformat": "^3.0.3",
+    "ejs": "^2.5.6",
+    "express": "^4.16.3",
+    "express-session": "^1.15.6",
+    "http": "0.0.0",
+    "jquery-validation": "^1.17.0",
+    "jsdom-global": "^3.0.2",
+    "merge-descriptors": "^1.0.1",
+    "mongodb": "^3.0.1",
+    "mongoose": "^4.10.8",
+    "mongoose-auto-increment": "^5.0.1",
+    "pdfmake": "^0.1.38",
+    "socket.io": "^2.0.4"
+  },
+  "devDependencies": {
+    "jsdom": "11.7.0",
+    "jsdom-global": "3.0.2",
+    "nodemon": "^1.17.5"
+  }
+}
