+++ server/server.js
@@ -1,67 +1,71 @@
+// *******
+// ALL SERVER CODE MOVED TO VMCAuction/server.js
+// *******
 
-var express = require("express");
-var app = express();
-var session = require('express-session');
-var secret = require('./config/secret.json')
-require('jsdom-global');
-require("pdfmake/build/pdfmake.js");
-require("pdfmake/build/vfs_fonts.js");
+// var express = require("express");
+// var app = express();
+// var session = require('express-session');
+// var secret = require('./config/secret.json')
+// require('jsdom-global');
+// require("pdfmake/build/pdfmake.js");
+// require("pdfmake/build/vfs_fonts.js");
 
-var mongoose = require('mongoose'),
-	Schema = mongoose.Schema,
-	autoIncrement = require('mongoose-auto-increment');
-mongoose.Promise = global.Promise;
+// var mongoose = require('mongoose'),
+// 	Schema = mongoose.Schema,
+// 	autoIncrement = require('mongoose-auto-increment');
+// mongoose.Promise = global.Promise;
 
 
-// Test change for github branch serverjs
+// // Test change for github branch serverjs
 
-app.use(session({
-  secret: secret.secret,
-  resave: false,
-  saveUninitialized: true,
-  rolling: true
-	//resets session timeout everytime the user interacts with the site
-}));
+// app.use(session({
+//   secret: secret.secret,
+//   resave: false,
+//   saveUninitialized: true,
+//   rolling: true
+// 	//resets session timeout everytime the user interacts with the site
+// }));
 
-// app.on("Category initializer", function(){
-// 	console.log("entered category initializer")
-// 	if(Category.checkIfEmpty() === true){
-// 		console.log("checkIfEmpty ran")
-// 		Category.initialize()
-// 	}
+// // app.on("Category initializer", function(){
+// // 	console.log("entered category initializer")
+// // 	if(Category.checkIfEmpty() === true){
+// // 		console.log("checkIfEmpty ran")
+// // 		Category.initialize()
+// // 	}
+// // })
+
+// var bodyParser = require('body-parser');
+// app.use(bodyParser.json());
+// app.use(bodyParser.urlencoded({ extended: true }));
+
+
+// var path = require("path");
+// app.use(express.static("../public"));
+
+// if (process.env.NODE_ENV === "production") {
+// 		app.use(express.static(path.join(__dirname, "/../wireframe")));
+// }else{
+// 		app.use(express.static(path.join(__dirname, "/../wireframe")));
+// }
+
+// app.set('views', path.join(__dirname, '../wireframe'));
+// app.set('view engine', 'ejs');
+
+// require('./config/mongoose.js');
+// require('./config/initialize.js');
+
+// var routesSetter = require('./config/routes.js');
+// routesSetter(app);
+
+
+// var port = process.env.PORT || 8000;
+// var server = app.listen(port, function() {
+//  console.log("listening on port "+port);
 // })
 
-var bodyParser = require('body-parser');
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({ extended: true }));
 
 
-var path = require("path");
-app.use(express.static("../public"));
-
-if (process.env.NODE_ENV === "production") {
-		app.use(express.static(path.join(__dirname, "/../wireframe")));
-}else{
-		app.use(express.static(path.join(__dirname, "/../wireframe")));
-}
-
-app.set('views', path.join(__dirname, '../wireframe'));
-app.set('view engine', 'ejs');
-
-require('./config/mongoose.js');
-require('./config/initialize.js');
-
-var routesSetter = require('./config/routes.js');
-routesSetter(app);
-
-
-var port = process.env.PORT || 8000;
-var server = app.listen(port, function() {
- console.log("listening on port "+port);
-})
-
-
-
+// ALL SOCKETS CODE LEFT ALONE
 
 /////////////// SOCKETS /////////////////
 var Package = require('./models/package.js')
